refactor: 카테고리 변경 로직 훅으로 분리
