+++ src/app/(main)/_components/molecules/CategoryButtonList.tsx
@@ -1,77 +1,17 @@
 'use client';
 
-import React, { useCallback, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import Swiper from 'swiper';
-import { usePathname, useRouter, useSearchParams } from 'next/navigation';
-import { useSearchStore } from '@/store/search';
-import { useCreateAgora } from '@/store/create';
-import { useShallow } from 'zustand/react/shallow';
-import { homeSegmentKey } from '@/constants/segmentKey';
 
 // import 'swiper/css';
 import 'swiper/css/free-mode';
 import 'swiper/css/mousewheel';
-import { isValidCategoryKey } from '@/utils/validation/validateCategoryKey';
 import { AGORACATEGORY } from '@/constants/consts';
-import isNull from '@/utils/validation/validateIsNull';
+import { useCategoryStatus } from '@/hooks/useCategoryStatus';
 import CategoryButton from '../atoms/CategoryButton';
 
 export default function CategoryButtonList() {
-  const { reset } = useSearchStore(
-    useShallow((state) => ({
-      reset: state.reset,
-    })),
-  );
-
-  const searchParams = useSearchParams();
-  const getCategorySearchParams = (): keyof typeof AGORACATEGORY => {
-    const categorySearchParams = searchParams.get('category');
-
-    if (isNull(categorySearchParams)) return '1';
-
-    if (categorySearchParams in AGORACATEGORY) {
-      return categorySearchParams as keyof typeof AGORACATEGORY;
-    }
-
-    return '1';
-  };
-
-  const pathname = usePathname();
-  const router = useRouter();
-  const { setCategory, category: selectedCategory } = useCreateAgora(
-    useShallow((state) => ({
-      setCategory: state.setCategory,
-      category: state.category,
-    })),
-  );
-
-  useEffect(() => {
-    setCategory(getCategorySearchParams());
-
-    return () => {
-      setCategory('1');
-    };
-  }, []);
-
-  const changeCategoryParams = useCallback(
-    (id: string) => {
-      if (pathname !== homeSegmentKey) return;
-
-      const newSearchParams = new URLSearchParams(searchParams);
-
-      newSearchParams.set('category', isValidCategoryKey(id) ? id : '1');
-      newSearchParams.delete('q');
-      reset();
-
-      const newUrl = `${homeSegmentKey}?${newSearchParams.toString()}`;
-      window.history.pushState(
-        { ...window.history.state, as: newUrl, url: newUrl },
-        '',
-        newUrl,
-      );
-    },
-    [router, searchParams, pathname, selectedCategory],
-  );
+  const { category: selectedCategory, setCategory } = useCategoryStatus();
 
   useEffect(() => {
     const swiper = new Swiper('.swiper', {
@@ -94,9 +34,9 @@ export default function CategoryButtonList() {
     };
   }, []);
 
-  useEffect(() => {
-    changeCategoryParams(selectedCategory);
-  }, [selectedCategory, changeCategoryParams]);
+  const handleClickCategoryBtn = (value: keyof typeof AGORACATEGORY) => {
+    setCategory(value);
+  };
 
   return (
     <div className="w-full mt-10 mb-0 pb-0 pl-0.5rem pr-0.5rem flex text-nowrap overflow-hidden ml-5">
@@ -106,7 +46,7 @@ export default function CategoryButtonList() {
             <button
               type="button"
               aria-label={`${category.innerText} 카테고리`}
-              onClick={() => setCategory(category.value)}
+              onClick={() => handleClickCategoryBtn(category.value)}
               className="swiper-slide"
               key={category.innerText}
             >
